split models chase and step
